+++ cmd/passphrases.go
@@ -37,3 +37,13 @@ func getPassphrase() string {
 	assert(len(passphrases) == 1, "multiple passphrases supplied; cannot continue")
 	return passphrases[0]
 }
+
+// getOptionalPassphrase fetches the passphrase if supplied by the user.
+func getOptionalPassphrase() string {
+	passphrases := getPassphrases()
+	if len(passphrases) == 0 {
+		return ""
+	}
+	assert(len(passphrases) == 1, "multiple passphrases supplied; cannot continue")
+	return passphrases[0]
+}