+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+Development:
+  - "account info" with verbose option now displays participants for distributed accounts
+  - fix issue where distributed account generation without a passphrase was not allowed
+
 1.5.8:
   - allow raw deposit transactions to be supplied to "deposit verify"
   - move functionality of "account withdrawalcredentials" to be part of "account info"